backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytesseract==0.3.10
Pillow==10.0.0
openpyxl==3.1.2
pdf2image==1.16.3
PyPDF2==3.0.1
regex==2023.10.3
Werkzeug==2.3.7
